Ctrl+D support, MSAA filtering, add/remove fields for entities, new undo-redo system, mac os support!

+++ NextCore/include/core/memory/linear_allocator.h
@@ -1,10 +1,12 @@
 #pragma once
 
-#include "core\memory\allocator.h"
+
+#include "core/memory/allocator.h"
+#include <new>
 
 struct LinearAllocator : Allocator {
-	size_t occupied;
-	size_t max_size;
+	u64 occupied;
+	u64 max_size;
 
 	char* memory;
 	Allocator* parent;
